Fix up spoof user redirection

+++ public/header.js
@@ -14,8 +14,8 @@ if (!localStorage.getItem('name')) {
     fetch(`${base}users`)
       .then((res) => res.json())
       .then((body) => {
-        if (!localStorage.getItem('name') || localStorage.getItem('google')) return;
-        if (!body.users.find(user => user[0] === localStorage.getItem('name') || user[1] === localStorage.getItem('id'))) {
+        if (!localStorage.getItem('name')) return;
+        if (!body.users.find(user => user[0] === localStorage.getItem('name') && user[1] === localStorage.getItem('id'))) {
           localStorage.clear();
           location.reload();
         }
@@ -24,8 +24,8 @@ if (!localStorage.getItem('name')) {
   fetch(`${base}users`)
     .then((res) => res.json())
     .then((body) => {
-      if (!localStorage.getItem('name') || localStorage.getItem('google')) return;
-      if (!body.users.find(user => user[0] === localStorage.getItem('name') || user[1] === localStorage.getItem('id'))) {
+      if (!localStorage.getItem('name')) return;
+      if (!body.users.find(user => user[0] === localStorage.getItem('name') && user[1] === localStorage.getItem('id'))) {
         localStorage.clear();
         location.reload();
       }
